exception propogation

+++ backend/TimeTracker/src/main/java/com/newput/rest/resource/EmpController.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
 import com.newput.domain.Employee;
+import com.newput.domain.TrackerException;
 import com.newput.service.EmpService;
 import com.newput.service.LoginService;
 import com.newput.service.SelectiveExcel;
@@ -66,40 +67,49 @@ public class EmpController {
 	private ExcelTimeSheet excelTimeSheet;
 
 	/**
+	 * @throws TrackerException
 	 * @POST Description : Use to add new user into the system and send the
-	 *       validation email to the registered mail id
-	 *       {@link EMailSender}
+	 *       validation email to the registered mail id {@link EMailSender}
 	 */
 	@Path("/register")
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
+	// @ExceptionHandler({TrackerException.class})
 	public JSONObject registerUser(@FormParam("firstName") String firstName, @FormParam("lastName") String lastName,
 			@FormParam("email") String email, @FormParam("dob") String dob, @FormParam("doj") String doj,
 			@FormParam("address") String address, @FormParam("contact") String contact,
 			@FormParam("gender") String gender, @FormParam("password") String password) {
-
-		String token = util.generateRandomString();
-		reqParser.setEmployeeValue(firstName, lastName, email, dob, doj, address, contact, gender, password, token);
-		empService.addUser(emp);
-		if (jsonResService.isSuccess()) {
-			emailSend.sendMail("registration");
+		try {
+			String token = util.generateRandomString();
+			reqParser.setEmployeeValue(firstName, lastName, email, dob, doj, address, contact, gender, password, token);
+			empService.addUser(emp);
+			if (jsonResService.isSuccess()) {
+				emailSend.sendMail("registration");
+			}
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Email id already registered").getMessage());
 		}
 		return jsonResService.responseSender();
+
 	}
 
 	@Path("/verify")
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject mailVerification(@FormParam("email") String emailId, @FormParam("token") String token) {
-		if (emailId != null && !emailId.equalsIgnoreCase("")) {
-			if (token != null && !token.equalsIgnoreCase("")) {
-				reqParser.setValidationValue(emailId, token);
-				empService.mailVerify(emp);
+		try {
+			if (emailId != null && !emailId.equalsIgnoreCase("")) {
+				if (token != null && !token.equalsIgnoreCase("")) {
+					reqParser.setValidationValue(emailId, token);
+					empService.mailVerify(emp);
+				} else {
+					jsonResService.errorResponse("token can not be blank");
+				}
 			} else {
-				jsonResService.errorResponse("token can not be blank");
+				jsonResService.errorResponse("Mail id can not be null");
 			}
-		} else {
-			jsonResService.errorResponse("Mail id can not be null");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Internal Server Error").getMessage());
 		}
 		return jsonResService.responseSender();
 	}
@@ -108,15 +118,19 @@ public class EmpController {
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject login(@FormParam("email") String email, @FormParam("password") String password) {
-		if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
-			if (password != null && !password.equalsIgnoreCase("")) {
-				reqParser.setSessionValue(email, password, "");
-				loginService.createSession(emp);
+		try {
+			if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
+				if (password != null && !password.equalsIgnoreCase("")) {
+					reqParser.setSessionValue(email, password, "");
+					loginService.createSession(emp);
+				} else {
+					jsonResService.errorResponse("password can not be blank");
+				}
 			} else {
-				jsonResService.errorResponse("password can not be blank");
+				jsonResService.errorResponse("Mail id can not be null and in proper format");
 			}
-		} else {
-			jsonResService.errorResponse("Mail id can not be null and in proper format");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Internal Server Error ").getMessage());
 		}
 		return jsonResService.responseSender();
 	}
@@ -129,23 +143,25 @@ public class EmpController {
 			@FormParam("lunchOut") String lunchOut, @FormParam("nightIn") String nightIn,
 			@FormParam("nightOut") String nightOut, @FormParam("workDesc") String workDesc,
 			@FormParam("empId") String emp_id) {
-		int id = Integer.parseInt(emp_id);
-		if (workdate != null && !workdate.equalsIgnoreCase("")) {
-			if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
-
-				timeSchedual.timeSheetValue(lunchIn, in, out, workdate, lunchOut, nightIn, nightOut, id);
-				if (workDesc != null && !workDesc.equalsIgnoreCase("")) {
-					reqParser.setDateSheetValue(workDesc, workdate, id);
-					timeSchedual.dateSheetValue();
+		try {
+			if (workdate != null && !workdate.equalsIgnoreCase("")) {
+				if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
+					int id = Integer.parseInt(emp_id);
+					timeSchedual.timeSheetValue(lunchIn, in, out, workdate, lunchOut, nightIn, nightOut, id);
+					if (workDesc != null && !workDesc.equalsIgnoreCase("")) {
+						reqParser.setDateSheetValue(workDesc, workdate, id);
+						timeSchedual.dateSheetValue();
+					}
+					timeSchedual.clearMap();
+				} else {
+					jsonResService.errorResponse("emp_id can not be null");
 				}
-				timeSchedual.clearMap();
 			} else {
-				jsonResService.errorResponse("emp_id can not be null");
+				jsonResService.errorResponse("Date can not be null");
 			}
-		} else {
-			jsonResService.errorResponse("Date can not be null");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user entry").getMessage());
 		}
-
 		return jsonResService.responseSender();
 	}
 
@@ -153,35 +169,47 @@ public class EmpController {
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject forgotPwd(@FormParam("email") String email) {
-
-		if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
-			String ptoken = util.generateRandomString();
-			empService.resetPassword(email, ptoken, "password");
-			if (jsonResService.isSuccess()) {
-				emailSend.sendMail("password");
+		try {
+			if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
+				String ptoken = util.generateRandomString();
+				empService.resetPassword(email, ptoken, "password");
+				if (jsonResService.isSuccess()) {
+					emailSend.sendMail("password");
+				}
+			} else {
+				jsonResService.errorResponse("Mail id can not be null and in proper format");
 			}
-		} else {
-			jsonResService.errorResponse("Mail id can not be null and in proper format");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Internal Server Error").getMessage());
 		}
 		return jsonResService.responseSender();
 	}
 
 	@Path("/excelExport")
 	@POST
-	@Produces("application/vnd.ms-excel")	
+	@Produces("application/vnd.ms-excel")
 	public Response excelExport(@FormParam("empId") String emp_id, @FormParam("month") String monthName,
 			@FormParam("year") String year) {
 		ResponseBuilder response = null;
-		if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
-			if (monthName != null && !monthName.equalsIgnoreCase("")) {
-				File file = excelTimeSheet.createExcelSheet(Integer.parseInt(emp_id), monthName, year);
-				response = Response.ok((Object) file);
-				response.header("Content-Disposition", "attachment; filename=time-sheet.xls");								
+		try {
+			if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
+				if (monthName != null && !monthName.equalsIgnoreCase("")) {
+					if (util.validCheck(monthName, year)) {
+						File file = excelTimeSheet.createExcelSheet(Integer.parseInt(emp_id), monthName, year);
+						response = Response.ok((Object) file);
+						response.header("Content-Disposition", "attachment; filename=time-sheet.xls");
+					} else {
+						jsonResService.errorResponse("Record is not avail before November 2015");
+					}
+				} else {
+					jsonResService.errorResponse("Please provide the month to select data");
+				}
+
 			} else {
-				jsonResService.errorResponse("Please provide the month to select data");
+				jsonResService.errorResponse("Please provide employee id to select data");
 			}
-		} else {
-			jsonResService.errorResponse("Please provide employee id to select data");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user").getMessage());
 		}
 		return response.build();
 	}
@@ -191,20 +219,24 @@ public class EmpController {
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject passwordVerification(@FormParam("empId") int empId, @FormParam("pToken") String pToken,
 			@FormParam("newPassword") String newPwd) {
-		if (empId > 0) {
-			if (pToken != null && !pToken.equalsIgnoreCase("")) {
-				if (newPwd != null && !newPwd.equalsIgnoreCase("")) {
-					newPwd = util.md5(newPwd);
-					reqParser.setPValidationValue(empId, pToken, newPwd);
-					empService.pwdVerify(emp);
+		try {
+			if (empId > 0) {
+				if (pToken != null && !pToken.equalsIgnoreCase("")) {
+					if (newPwd != null && !newPwd.equalsIgnoreCase("")) {
+						newPwd = util.md5(newPwd);
+						reqParser.setPValidationValue(empId, pToken, newPwd);
+						empService.pwdVerify(emp);
+					} else {
+						jsonResService.errorResponse("Password can not be blank");
+					}
 				} else {
-					jsonResService.errorResponse("Password can not be blank");
+					jsonResService.errorResponse("token can not be blank");
 				}
 			} else {
-				jsonResService.errorResponse("token can not be blank");
+				jsonResService.errorResponse("Mail id can not be null");
 			}
-		} else {
-			jsonResService.errorResponse("Mail id can not be null");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user").getMessage());
 		}
 		return jsonResService.responseSender();
 	}
@@ -214,16 +246,23 @@ public class EmpController {
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject monthlyExcel(@FormParam("month") String monthName, @FormParam("empId") String emp_id,
 			@FormParam("year") String year) {
-		if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
-			if (monthName != null && !monthName.equalsIgnoreCase("")) {
-				excel.monthSheet(monthName, Integer.parseInt(emp_id), year);
+		try {
+			if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
+				if (monthName != null && !monthName.equalsIgnoreCase("")) {
+					if (util.validCheck(monthName, year)) {
+						excel.monthSheet(monthName, Integer.parseInt(emp_id), year);
+					} else {
+						jsonResService.errorResponse("Record is not avail before November 2015");
+					}
+				} else {
+					jsonResService.errorResponse("Please provide the month to select data");
+				}
 			} else {
-				jsonResService.errorResponse("Please provide the month to select data");
+				jsonResService.errorResponse("Please provide employee id to select data");
 			}
-		} else {
-			jsonResService.errorResponse("Please provide employee id to select data");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user").getMessage());
 		}
-
 		return jsonResService.responseSender();
 	}
 
@@ -232,18 +271,26 @@ public class EmpController {
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject mailExcelSheet(@FormParam("empId") String emp_id, @FormParam("month") String monthName,
 			@FormParam("year") String year) {
-		if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
-			if (monthName != null && !monthName.equalsIgnoreCase("")) {
-				File file = excelTimeSheet.createExcelSheet(Integer.parseInt(emp_id), monthName, year);
-				if (jsonResService.isSuccess()) {
-					emailSend.sendExcelSheet(excelTimeSheet.getEmpEmail(Integer.parseInt(emp_id)), file);
-					file.delete();
+		try {
+			if (emp_id != null && !emp_id.equalsIgnoreCase("")) {
+				if (monthName != null && !monthName.equalsIgnoreCase("")) {
+					if (util.validCheck(monthName, year)) {
+						File file = excelTimeSheet.createExcelSheet(Integer.parseInt(emp_id), monthName, year);
+						if (jsonResService.isSuccess()) {
+							emailSend.sendExcelSheet(excelTimeSheet.getEmpEmail(Integer.parseInt(emp_id)), file);
+							file.delete();
+						}
+					} else {
+						jsonResService.errorResponse("Record is not avail before November 2015");
+					}
+				} else {
+					jsonResService.errorResponse("Please provide the month to select data");
 				}
 			} else {
-				jsonResService.errorResponse("Please provide the month to select data");
+				jsonResService.errorResponse("Please provide employee id to select data");
 			}
-		} else {
-			jsonResService.errorResponse("Please provide employee id to select data");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user").getMessage());
 		}
 		return jsonResService.responseSender();
 	}
@@ -252,27 +299,36 @@ public class EmpController {
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject resendMail(@FormParam("email") String email, @FormParam("flag") String flag) {
-		if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
-			if (flag != null && !flag.equalsIgnoreCase("")) {
-				String token = util.generateRandomString();
-				empService.resetPassword(email, token,flag);
-				if (jsonResService.isSuccess()) {
-					emailSend.sendMail(flag);
+		try {
+			if (email != null && !email.equalsIgnoreCase("") && util.mailFormat(email)) {
+				if (flag != null && !flag.equalsIgnoreCase("")) {
+					String token = util.generateRandomString();
+					empService.resetPassword(email, token, flag);
+					if (jsonResService.isSuccess()) {
+						emailSend.sendMail(flag);
+					}
+				} else {
+					jsonResService.errorResponse("Flag is must");
 				}
 			} else {
-				jsonResService.errorResponse("Flag is must");
+				jsonResService.errorResponse("Email id not valid");
 			}
-		} else {
-			jsonResService.errorResponse("Email id not valid");
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Internal Server Error ").getMessage());
 		}
 		return jsonResService.responseSender();
+
 	}
 
 	@Path("/signOut")
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	public JSONObject signOut(@FormParam("empId") String emp_id) {
-		loginService.signOut(Integer.parseInt(emp_id));
+		try {
+			loginService.signOut(Integer.parseInt(emp_id));
+		} catch (Exception ex) {
+			jsonResService.errorResponse(new TrackerException("Invalid user").getMessage());
+		}
 		return jsonResService.responseSender();
 	}
